Trigger an event for each notable character, rather than for each span
Instead of firing a pair of events around each span of characters that
represents a single entity (which is difficult, and requires maintaining
state in BaseScanner and LineEventHandler), we can simply fire a
'comment' or 'code' event for each such character we see.

BaseScanner keeps a very simple state: where did the current line begin,
and have comments or code been seen in this line? All the logic
collapses into a few lines in notifyNewline().

In this patch, LineEvenHandler is not used at all. We simply print
directly to stdout from notifyNewline(). In actual production code, the
scanner could accept a callback which could do the printing, or save to
the DB, or whatever.