Adds (failing) tests to cover unterminated multiline strings, comments
As currently defined, unterminated multi-line strings and comments will
crash our scanners. This is because when the string or comment fails to
terminate, at EOF Ragel will backtrack all the way to the beginning of
the string or comment, and re-scan them against a different pattern.

That's bad, because we've already committed to a line count for all of
that content, and can't count it again.

The fix requires us to re-design our state machines such that once a
newline has been consumed, it is not possible to backtrack.

A fix is forthcoming. For now, the tests fail.