Adds file counts to line count summary report
FileFinder has some compiler warnings related to generics that I don't quite
understand.

For the walk() method, The compiler warns that Collection is a raw type,
and should be parameterized.

If I try to paramterize it, then I fail to match the signature of the
method I am trying to override, which leads to a compiler error.

Examples on the web don't seem to discuss this issue. Hmm.