Another giant refactor: Allow languages to share scanners
There's no reason to cut and paste the exact same C scanner over and
over to make a dozen scanners that are different in name only.

Even with the 'c.rl' include, this is a lot of wasted effort.

I've changed things so that a language can declare which scanner it
wants to use, and any language can use any scanner.

This removes the need for the abstract getLanguage() method in all of
the ragel.

It also means that detection now returns a Language, and a Language
can be asked to instantiate a Scanner.

This simplifies a lot of things. It is now possible to add a new
language by editing Language.java *only*, if you are sharing an existing
scanner.