Resolvers are now properly invoked for ambiguous extensions.
Also, a major change: OhcountException is now a RuntimeException.

I was motiviated to do this because all of the Detector methods are now
capable of throwing OhcountException, thanks to dynamic lookup of
Resolvers. This then leads to OhcountException propagating into every
method in the application.

Changing to RuntimeException is probably correct. We throw
OhcountException only for fatal errors: when classes can't be
instantiated, or libraries can't be loaded. There's no meaningful way to
continue running after this happens.