Initial Fortran Resolver
Because so many different extensions must be disambiguated, I chose not
to create a Resolver for each extension. Instead, I've special-cased the
use of FortranResolver. I don't like that this was required. Smalltalk's
*.st resolver also had to use a hack. Perhaps we need to rethink how
Resolvers are invoked.

Also, disambiguate_fortran() in the original Ohcount is much more
complex than what I have here. Unfortunatley, that code is (a) hard to
understand, and (b) doesn't have enough test coverage. After porting
only a fraction of the old disambiguator, I was able to pass all of the
original Ohcount test cases.

I will continue to try to decipher what rules the old disambiguator is
enforcing; for now, this gets us 90% there.